Extracting eligible method definitions for IL rewrite
Originally this was implemented using the .net builtin reflection api,
but if take that approach, that mean after getting a list of methods
to be il processed, will still have to search each method in cecil
so il can be weaved in, that is O(n^2) at best. So this is the attempt
to reimplement that using cecil directly, this commit is for checking
if aspect is applied on the assembly level, if so, include all methods
from all types as eligible.