More on the Around aspect
Still some bugs/concerns to address. For ex, when generating a replacement
function the fields inside the Around aspect need to be copied and
instantiated as appropriate. Current this setp is missing so everything
has to be contained inside the Invoke() function for it to work.

Another thing is, when compiled in Release configuration the weaving
errors out, need to investigate it furtuer.