Importing types inorder to write out the assembly
If the Around aspect is defined in a different assembly we need to modify
that assembly to change the IL instruction. In order to get around the
'member x declared in another assembly must be imported' problem currently
I'm importing bunch of types accordingly. Preliminary test shows ok, but
need more testing.